Changed Int to Float

+++ app/src/main/java/lu/hao/cryptowatchers/CoinHistory.kt
@@ -2,6 +2,6 @@ package lu.hao.cryptowatchers
 
 import com.google.gson.annotations.SerializedName
 
-data class CoinHistory(@SerializedName("market_cap") var marketCap: List<List<Int>>? = null,
-                       var price: List<List<Int>>? = null,
-                       var volume: List<List<Int>>? = null)
+data class CoinHistory(@SerializedName("market_cap") var marketCap: List<List<Float>>? = null,
+                       var price: List<List<Float>>? = null,
+                       var volume: List<List<Float>>? = null)
